application/backend/requirements.txt
Flask
Flask-CORS
Click
python-dotenv
bcrypt
mysql-connector-python 
didkit
PyJWT
asgiref
uvicorn